+++ src/Readable/Frame.php
@@ -1,0 +1,44 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\IO\Readable;
+
+use Innmind\Immutable\{
+    Maybe,
+    Str,
+};
+
+/**
+ * @template T
+ */
+interface Frame
+{
+    /**
+     * @param callable(?positive-int): Maybe<Str> $read
+     * @param callable(): Maybe<Str> $readLine
+     *
+     * @return Maybe<T>
+     */
+    public function __invoke(
+        callable $read,
+        callable $readLine,
+    ): Maybe;
+
+    /**
+     * @template U
+     *
+     * @param callable(T): U $map
+     *
+     * @return self<U>
+     */
+    public function map(callable $map): self;
+
+    /**
+     * @template U
+     *
+     * @param callable(T): self<U> $map
+     *
+     * @return self<U>
+     */
+    public function flatMap(callable $map): self;
+}